+++ src/Infrastructure/Handler/IRepositoryCommand.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace ViajaNet.JobApplication.Infrastructure
+{
+    public interface IRepositoryCommand
+    {
+        Task<string> CreateAsync();
+    }
+}