+++ src/Infrastructure/Handler/ISqlServerService.cs
@@ -1,0 +1,6 @@
+namespace ViajaNet.JobApplication.Infrastructure
+{
+    public interface ISqlServerService
+    {
+    }
+}